# backend/internal/contracts/openapi.gen.go
// Package contracts provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package contracts

import (
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// Defines values for ProblemCategory.
const (
	Authentication ProblemCategory = "authentication"
	Authorization  ProblemCategory = "authorization"
	Conflict       ProblemCategory = "conflict"
	Internal       ProblemCategory = "internal"
	NotFound       ProblemCategory = "not_found"
	Validation     ProblemCategory = "validation"
)

// Valid indicates whether the value is a known member of the ProblemCategory enum.
func (e ProblemCategory) Valid() bool {
	switch e {
	case Authentication:
		return true
	case Authorization:
		return true
	case Conflict:
		return true
	case Internal:
		return true
	case NotFound:
		return true
	case Validation:
		return true
	default:
		return false
	}
}

// ContractProbeRequest defines model for ContractProbeRequest.
type ContractProbeRequest struct {
	Value string `json:"value"`
}

// ContractProbeResponse defines model for ContractProbeResponse.
type ContractProbeResponse struct {
	Value string `json:"value"`
}

// Problem defines model for Problem.
type Problem struct {
	Category ProblemCategory `json:"category"`
	Code     string          `json:"code"`

	// Detail Safe client-facing context without infrastructure details.
	Detail   *string `json:"detail,omitempty"`
	Instance *string `json:"instance,omitempty"`
	Status   int     `json:"status"`
	Title    string  `json:"title"`
	TraceId  string  `json:"traceId"`

	// Type Stable problem type identifier.
	Type       string                 `json:"type"`
	Violations *[]ValidationViolation `json:"violations,omitempty"`
}

// ProblemCategory defines model for Problem.Category.
type ProblemCategory string

// ValidationViolation defines model for ValidationViolation.
type ValidationViolation struct {
	Field   string  `json:"field"`
	Message *string `json:"message,omitempty"`
	Rule    string  `json:"rule"`
}

// PageNumber defines model for PageNumber.
type PageNumber = int

// PageSize defines model for PageSize.
type PageSize = int

// AuthenticationProblem defines model for AuthenticationProblem.
type AuthenticationProblem = Problem

// AuthorizationProblem defines model for AuthorizationProblem.
type AuthorizationProblem = Problem

// ConflictProblem defines model for ConflictProblem.
type ConflictProblem = Problem

// InternalProblem defines model for InternalProblem.
type InternalProblem = Problem

// NotFoundProblem defines model for NotFoundProblem.
type NotFoundProblem = Problem

// ValidationProblem defines model for ValidationProblem.
type ValidationProblem = Problem

// ProbeContractParams defines parameters for ProbeContract.
type ProbeContractParams struct {
	// Page One-based result page.
	Page *PageNumber `form:"page,omitempty" json:"page,omitempty"`

	// PageSize Maximum results returned per page.
	PageSize *PageSize `form:"pageSize,omitempty" json:"pageSize,omitempty"`
}

// ProbeContractJSONRequestBody defines body for ProbeContract for application/json ContentType.
type ProbeContractJSONRequestBody = ContractProbeRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// ProbeContract Validate generated transport conformance
	// (POST /contract/probe)
	ProbeContract(w http.ResponseWriter, r *http.Request, params ProbeContractParams)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// ProbeContract Validate generated transport conformance
// (POST /contract/probe)
func (_ Unimplemented) ProbeContract(w http.ResponseWriter, r *http.Request, params ProbeContractParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// ProbeContract operation middleware
func (siw *ServerInterfaceWrapper) ProbeContract(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// Parameter object where we will unmarshal all parameters from the context
	var params ProbeContractParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "page", r.URL.Query(), &params.Page, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "page"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		}
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "pageSize", r.URL.Query(), &params.PageSize, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		var requiredError *runtime.RequiredParameterError
		if errors.As(err, &requiredError) {
			siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "pageSize"})
		} else {
			siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		}
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ProbeContract(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/contract/probe", wrapper.ProbeContract)
	})

	return r
}

type AuthenticationProblemApplicationProblemPlusJSONResponse Problem

type AuthorizationProblemApplicationProblemPlusJSONResponse Problem

type ConflictProblemApplicationProblemPlusJSONResponse Problem

type InternalProblemApplicationProblemPlusJSONResponse Problem

type NotFoundProblemApplicationProblemPlusJSONResponse Problem

type ValidationProblemApplicationProblemPlusJSONResponse Problem

type ProbeContractRequestObject struct {
	Params ProbeContractParams
	Body   *ProbeContractJSONRequestBody
}

type ProbeContractResponseObject interface {
	VisitProbeContractResponse(w http.ResponseWriter) error
}

type ProbeContract200JSONResponse ContractProbeResponse

func (response ProbeContract200JSONResponse) VisitProbeContractResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ProbeContract400ApplicationProblemPlusJSONResponse struct {
	ValidationProblemApplicationProblemPlusJSONResponse
}

func (response ProbeContract400ApplicationProblemPlusJSONResponse) VisitProbeContractResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type ProbeContract401ApplicationProblemPlusJSONResponse struct {
	AuthenticationProblemApplicationProblemPlusJSONResponse
}

func (response ProbeContract401ApplicationProblemPlusJSONResponse) VisitProbeContractResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type ProbeContract403ApplicationProblemPlusJSONResponse struct {
	AuthorizationProblemApplicationProblemPlusJSONResponse
}

func (response ProbeContract403ApplicationProblemPlusJSONResponse) VisitProbeContractResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(403)
	_, err := buf.WriteTo(w)
	return err
}

type ProbeContract404ApplicationProblemPlusJSONResponse struct {
	NotFoundProblemApplicationProblemPlusJSONResponse
}

func (response ProbeContract404ApplicationProblemPlusJSONResponse) VisitProbeContractResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(404)
	_, err := buf.WriteTo(w)
	return err
}

type ProbeContract409ApplicationProblemPlusJSONResponse struct {
	ConflictProblemApplicationProblemPlusJSONResponse
}

func (response ProbeContract409ApplicationProblemPlusJSONResponse) VisitProbeContractResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type ProbeContract500ApplicationProblemPlusJSONResponse struct {
	InternalProblemApplicationProblemPlusJSONResponse
}

func (response ProbeContract500ApplicationProblemPlusJSONResponse) VisitProbeContractResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/problem+json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// ProbeContract Validate generated transport conformance
	// (POST /contract/probe)
	ProbeContract(ctx context.Context, request ProbeContractRequestObject) (ProbeContractResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	if options.RequestErrorHandlerFunc == nil {
		options.RequestErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	if options.ResponseErrorHandlerFunc == nil {
		options.ResponseErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		}
	}
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// ProbeContract operation middleware
func (sh *strictHandler) ProbeContract(w http.ResponseWriter, r *http.Request, params ProbeContractParams) {
	var request ProbeContractRequestObject

	request.Params = params

	var body ProbeContractJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.ProbeContract(ctx, request.(ProbeContractRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "ProbeContract")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ProbeContractResponseObject); ok {
		if err := validResponse.VisitProbeContractResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}